requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
oso==0.27.3
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
aryn-sdk==0.1.10
llama-index==0.12.2
llama-index-vector-stores-pinecone==0.4.1
python-multipart==0.0.20
openai==1.58.1
pinecone-client==4.1.2
python-jose[cryptography]
python-dotenv